refactor: move journals into own files
