+++ packages/core/test/new-api/execution/execution-engine.ts
@@ -5,8 +5,8 @@ import { defineModule } from "../../../src/new-api/define-module";
 import { Batcher } from "../../../src/new-api/internal/batcher";
 import { ExecutionEngine } from "../../../src/new-api/internal/execution/execution-engine";
 import { BasicExecutionStrategy } from "../../../src/new-api/internal/execution/execution-strategy";
+import { MemoryJournal } from "../../../src/new-api/internal/journal/memory-journal";
 import { ModuleConstructor } from "../../../src/new-api/internal/module-builder";
-import { MemoryJournal } from "../../../src/new-api/journal";
 import {
   Journal,
   JournalableMessage,
